feature: WEOS-1343 As a developer I should be able to setup an OpenID Connect security scheme -Changed the openid middleware name to OpenIDMiddleware -Add a check in the middleware to make sure the security scheme used has type openidconnect -Moved the getjwlurl to the utils file -Updated the global intializer to check the security schem type and attach the openidmiddleware once the type is openidconnect
