+++ controllers/rest/api.go
@@ -332,7 +332,7 @@ func (p *RESTAPI) Initialize(ctxt context.Context) error {
 
 	//register standard middleware
 	p.RegisterMiddleware("Context", Context)
-	p.RegisterMiddleware("AuthorizationMiddleware", AuthorizationMiddleware)
+	p.RegisterMiddleware("OpenIDMiddleware", OpenIDMiddleware)
 	p.RegisterMiddleware("CreateMiddleware", CreateMiddleware)
 	p.RegisterMiddleware("CreateBatchMiddleware", CreateBatchMiddleware)
 	p.RegisterMiddleware("UpdateMiddleware", UpdateMiddleware)
