+++ controllers/rest/global_initializers.go
@@ -9,8 +9,15 @@ import (
 //Security adds authorization middleware to the initialize context
 func Security(ctxt context.Context, api *RESTAPI, swagger *openapi3.Swagger) (context.Context, error) {
 	middlewares := GetOperationMiddlewares(ctxt)
-	if swagger.Security != nil && len(swagger.Security) != 0 {
-		if middleware, _ := api.GetMiddleware("AuthorizationMiddleware"); middleware != nil {
+	found := false
+	for _, scheme := range swagger.Components.SecuritySchemes {
+		//checks if the security scheme has type openIdConnect
+		if scheme.Value.Type == "openIdConnect" {
+			found = true
+		}
+	}
+	if found {
+		if middleware, _ := api.GetMiddleware("OpenIDMiddleware"); middleware != nil {
 			middlewares = append(middlewares, middleware)
 		}
 		ctxt = context.WithValue(ctxt, weosContext.MIDDLEWARES, middlewares)
